refactor(web): 리뷰 반영
