+++ apps/web/src/app/(auth)/on-boarding/[step]/components/client-footer/ClientFooter.tsx
@@ -5,7 +5,7 @@ import { BottomCTAButton } from '@snappin/design-system';
 import { ROUTES } from '@/constants/routes/routes';
 import { TOTAL_STEP_COUNT } from '@/app/(auth)/on-boarding/[step]/constants/onBoardingSteps';
 import type { OnBoardingStep } from '@/app/(auth)/on-boarding/[step]/types/onBoardingStep';
-import { useOnBoardingFormContext } from '@/app/(auth)/on-boarding/[step]/hooks/useOnBoardingFormContext';
+import { useOnBoardingFormContext } from '@/app/(auth)/on-boarding/hooks/useOnBoardingFormContext';
 
 type ClientFooterProps = {
   step: number;
