+++ apps/web/src/app/(auth)/on-boarding/layout.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react';
 import ClientNavigation from '@/app/(auth)/on-boarding/components/client-navigation/ClientNavigation';
-import { OnBoardingFormProvider } from '@/app/(auth)/on-boarding/[step]/hooks/useOnBoardingFormContext';
+import { OnBoardingFormProvider } from '@/app/(auth)/on-boarding/hooks/useOnBoardingFormContext';
 
 export default function Layout({ children }: { children: ReactNode }) {
   return (
